Notification and swipe

+++ app/src/main/java/it/appoinkments/data/Appointment.kt
@@ -8,7 +8,7 @@ import java.util.*
 @Entity(tableName = "appointment")
 @TypeConverters(DateConverter::class, AppointmentTypeConverter::class)
 data class Appointment (@PrimaryKey(autoGenerate = true) val aid: Int = 0,
-                        var load_id: Int,
+                        var load_id: Long,
                         var date: Date,
                         var type: AppointmentType,
                         var farmer: String,
